# social-media/backend/docs/swagger-auth/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.EngagementData:
    properties:
      comments:
        type: integer
      date:
        type: string
      likes:
        type: integer
      shares:
        type: integer
      views:
        type: integer
    type: object
...
  internal_handlers.FollowerGrowth:
    properties:
      count:
        type: integer
      date:
        type: string
      lost:
        type: integer
      new:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Authentication service for SocialHub platform
  title: SocialHub Auth API
  version: 1.0.0
paths:
...
  /api/v1/analytics/engagement:
    get:
      description: Get user engagement metrics over time
      parameters:
      - default: 7d
        description: Time period (7d, 30d, 90d)
        in: query
        name: period
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.EngagementData'
            type: array
      security:
      - BearerAuth: []
      summary: Get engagement analytics
      tags:
      - analytics
  /api/v1/analytics/followers:
    get:
      description: Get user follower growth over time
      parameters:
      - default: 7d
        description: Time period (7d, 30d, 90d)
        in: query
        name: period
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.FollowerGrowth'
            type: array
      security:
      - BearerAuth: []
      summary: Get follower growth analytics
      tags:
      - analytics
  /api/v1/analytics/stats:
    get:
      description: Get overall user statistics
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user statistics
      tags:
      - analytics
...
swagger: "2.0"